chore: move animation types into animations.ts
chore: move util animation types into utils.ts
