--- src/types/animations/keyframes.type.ts
@@ -1,9 +1,0 @@
-type Axis = "x" | "y"
-
-export interface GenerateXYKeyframeProps {
-  axis: Axis
-  startDistance: number
-  endDistance: number
-}
-
-export type Keyframe = Record<string, Record<string, string>>
